requirements.txt
annotated-types~=0.7.0
anyio~=4.9.0
certifi~=2025.6.15
click~=8.2.1
dnspython~=2.7.0
email_validator~=2.2.0
fastapi~=0.115.14
fastapi-cli~=0.0.7
h11~=0.16.0
httpcore~=1.0.9
httptools~=0.6.4
httpx~=0.28.1
idna~=3.10
itsdangerous~=2.2.0
Jinja2~=3.1.6
markdown-it-py~=3.0.0
MarkupSafe~=3.0.2
mdurl~=0.1.2
orjson~=3.10.18
pydantic~=2.11.7
pydantic-extra-types~=2.10.5
pydantic-settings~=2.10.1
pydantic_core~=2.33.2
Pygments~=2.19.2
python-dotenv~=1.1.1
python-multipart~=0.0.20
PyYAML~=6.0.2
rich~=14.0.0
rich-toolkit~=0.14.7
shellingham~=1.5.4
sniffio~=1.3.1
starlette~=0.46.2
typer~=0.16.0
typing-inspection~=0.4.1
typing_extensions~=4.14.0
ujson~=5.10.0
uvicorn~=0.35.0
uvloop~=0.21.0
watchfiles~=1.1.0
websockets~=15.0.1
sqlalchemy~=2.0.41
asyncpg~=0.30.0
alembic~=1.16.2
psycopg2-binary~=2.9.10
pytest~=8.4.1
pytest-asyncio~=1.0.0
cohere~=5.16.1
pandas~=2.3.1
numpy~=2.3.2
scikit-learn~=1.7.1
xgboost~=3.1.1
greenlet~=3.2.4